requirements.txt
# Core deep-learning dependencies
torch>=2.6.0
torchvision>=0.21.0
ultralytics>=8.3.0

# SAHI – Slicing Aided Hyper Inference
sahi>=0.11.19

# Computer-vision / image processing
opencv-python>=4.10.0
Pillow>=10.4.0
albumentations>=1.4.21

# Data utilities
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.13.0
matplotlib>=3.9.0
seaborn>=0.13.0
tqdm>=4.66.0
PyYAML>=6.0

# Deduplication
imagehash>=4.3.1

# Alert & notification
python-telegram-bot>=21.6
twilio>=9.3.5
playsound>=1.3.0
requests>=2.32.0

# Location / mapping
# vietmap-sdk is fetched via REST API – no dedicated package needed

# Utilities
python-dotenv>=1.0.0
click>=8.1.0

# Testing
pytest>=8.3.0
pytest-cov>=5.0.0
